third_party/blink/web_tests/fast/events/drag-and-drop-expected.txt

This test checks that drag-and-drop support works and conforms to the HTML 5 specification.
For each effectAllowed, iterates through the possible dropEffects: none, copy, move, link, dummy.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE


When effectAllowed == "uninitialized"

PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "copy"
PASS event.dataTransfer.dropEffect is "move"
PASS event.dataTransfer.dropEffect is "link"
PASS event.dataTransfer.dropEffect is "copy"

When effectAllowed == "undefined"

PASS event.dataTransfer.effectAllowed is "uninitialized"
PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.effectAllowed is "uninitialized"
PASS event.dataTransfer.dropEffect is "copy"
PASS event.dataTransfer.effectAllowed is "uninitialized"
PASS event.dataTransfer.dropEffect is "move"
PASS event.dataTransfer.effectAllowed is "uninitialized"
PASS event.dataTransfer.dropEffect is "link"
PASS event.dataTransfer.effectAllowed is "uninitialized"
PASS event.dataTransfer.dropEffect is "copy"

When effectAllowed == "none"

PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "none"

When effectAllowed == "all"

PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "copy"
PASS event.dataTransfer.dropEffect is "move"
PASS event.dataTransfer.dropEffect is "link"
PASS event.dataTransfer.dropEffect is "copy"

When effectAllowed == "copy"

PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "copy"
PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "copy"

When effectAllowed == "move"

PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "move"
PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "move"

When effectAllowed == "link"

PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "link"
PASS event.dataTransfer.dropEffect is "link"

When effectAllowed == "copyMove"

PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "copy"
PASS event.dataTransfer.dropEffect is "move"
PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "move"

When effectAllowed == "copyLink"

PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "copy"
PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "link"
PASS event.dataTransfer.dropEffect is "copy"

When effectAllowed == "linkMove"

PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.dropEffect is "move"
PASS event.dataTransfer.dropEffect is "link"
PASS event.dataTransfer.dropEffect is "move"

When effectAllowed == "dummy"

PASS event.dataTransfer.effectAllowed is "uninitialized"
PASS event.dataTransfer.dropEffect is "none"
PASS event.dataTransfer.effectAllowed is "uninitialized"
PASS event.dataTransfer.dropEffect is "copy"
PASS event.dataTransfer.effectAllowed is "uninitialized"
PASS event.dataTransfer.dropEffect is "move"
PASS event.dataTransfer.effectAllowed is "uninitialized"
PASS event.dataTransfer.dropEffect is "link"
PASS event.dataTransfer.effectAllowed is "uninitialized"
PASS event.dataTransfer.dropEffect is "copy"

TEST COMPLETE
